// apps/web/src/sanity/sanity.types.ts
export type Navbar = {
  _id: string;
  _type: "navbar";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  links?: Array<{
    text?: string;
    link?: Link;
    sublinks?: Array<{
      text?: string;
      link?: Link;
      _key: string;
    }>;
    _key: string;
  }>;
};
// ...
export type Link = {
  _type: "link";
  text?: string;
  type?: string;
  internalLink?: {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "post";
  } | {
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    [internalGroqTypeReferenceTo]?: "author";
  };
  url?: string;
  email?: string;
  phone?: string;
  value?: string;
  blank?: boolean;
  parameters?: string;
  anchor?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;